# CUDA: add BF16 support