This PR adds BF16 support for CUDA/HIP. For large batch sizes the BF16 data is converted to FP32, then FP32 cuBLAS GEMM is used. It seems that cuBLAS unfortunately does not have support for BF16 tensor cores. For batch size 1 I added a template parameter to `mul_mat_vec` to specify the input type as either FP16 or BF16. The calculations are done using FP32 arithmetic since BF16 hardware support for operations other than matrix multiplications are only available with compute capability 9.0 and the highest that I own is 8.9. I will purchase a Blackwell GPU in a few weeks when they come out and revisit this.

Performance:

| Model           | CPU             | GPU        | FlashAttention     | Test     |     t/s master |     t/s cuda-bf16-3 |     Speedup |
| :-------------- | :-------------- | :--------- | :----------------- | :------- | -------------: | ------------------: | ----------: |
| gemma 2B BF16   | EPYC 7742       | RTX 4090   | No                 | pp512    |         565.29 |            12552.04 |       22.20 |
| gemma 2B BF16   | EPYC 7742       | RTX 4090   | No                 | tg128    |          22.47 |              158.23 |        7.04 |
| gemma 2B BF16   | EPYC 7742       | RTX 4090   | Yes                | pp512    |         564.51 |            12662.16 |       22.43 |
| gemma 2B BF16   | EPYC 7742       | RTX 4090   | Yes                | tg128    |          22.43 |              157.44 |        7.02 |
| gemma 2B BF16   | Ryzen 9 5950X   | RTX 3090   | No                 | pp512    |         256.16 |             5901.91 |       23.04 |
| gemma 2B BF16   | Ryzen 9 5950X   | RTX 3090   | No                 | tg128    |           8.01 |              142.12 |       17.73 |
| gemma 2B BF16   | Ryzen 9 5950X   | RTX 3090   | Yes                | pp512    |         267.93 |             5901.28 |       22.03 |
| gemma 2B BF16   | Ryzen 9 5950X   | RTX 3090   | Yes                | tg128    |           8.09 |              140.63 |       17.39 |
| gemma 2B BF16   | Xeon E5-2683 v4 | RX 6800    | No                 | pp512    |         130.89 |             1428.06 |       10.91 |
| gemma 2B BF16   | Xeon E5-2683 v4 | RX 6800    | No                 | tg128    |          10.19 |               53.77 |        5.28 |
| gemma 2B BF16   | Xeon E5-2683 v4 | RX 6800    | Yes                | pp512    |         128.57 |             1195.52 |        9.30 |
| gemma 2B BF16   | Xeon E5-2683 v4 | RX 6800    | Yes                | tg128    |          10.09 |               51.66 |        5.12 |
| gemma 2B BF16   | Xeon E5-2683 v4 | P40        | No                 | pp512    |         128.99 |             1625.54 |       12.60 |
| gemma 2B BF16   | Xeon E5-2683 v4 | P40        | No                 | tg128    |          10.80 |               55.97 |        5.18 |
| gemma 2B BF16   | Xeon E5-2683 v4 | P40        | Yes                | pp512    |         130.86 |             1476.22 |       11.28 |
| gemma 2B BF16   | Xeon E5-2683 v4 | P40        | Yes                | tg128    |          10.77 |               54.01 |        5.02 |