@sorasoras 2025-01-06
@JohannesGaessler  
I am bit confused. Doesn't 4090 support BF16 tensor? 
"BF16 hardware support for operations other than matrix multiplications are only available with compute capability 9.0"
what exact operation that support by blackwell but not ada?