@JohannesGaessler 2025-01-06
Addition, multiplication, division, etc.